sample generation, additional fixes
git-svn-id: http://svn.sgscript.org/tacstrike@246 89f9ca44-2ff2-4b64-a49d-2a4f1987ee72

+++ src/tools/rdgen.cpp
@@ -667,7 +667,7 @@ int main( int argc, char* argv[] )
 					printf( "Removing unused bone: \"%.*s\"\n", (int) B.name.size(), B.name.data() );
 					for( size_t j = i + 1; j < mesh_bones.size(); ++j )
 					{
-						if( mesh_bones[ j ].parent_id == i )
+						if( mesh_bones[ j ].parent_id == (int) i )
 							mesh_bones[ j ].parent_id = mesh_bones[ i ].parent_id;
 						else if( mesh_bones[ j ].parent_id > (int) i )
 							mesh_bones[ j ].parent_id--;
